MiGS: Update data hash 3DS fields
MiGS had existing support for passing 3DS auth fields, but it
seemed to be broken and not actually tested in the test suite.
Because the keys were strings and the rest of the keys were
symbols, there was an error when trying to do a sort on the keys.
This PR updates the 3ds keys to be symbols. Pending test
confirmation (current credentials resulting in a currency type
error)